+++ Specs/Model/ActivitySpec.swift
@@ -12,9 +12,8 @@ import Nimble
 class ActivitySpec: QuickSpec {
     override func spec() {
         it("converts Post activities from JSON") {
-            let (parsedActivity, parsedLinked) = stubbedJSONDataWithLinked("activity-own-post", "activity")
-            
-            let activity = Activity.fromJSON(parsedActivity, linked: parsedLinked) as Activity
+            let parsedActivity = stubbedJSONDataWithLinked("activity-own-post", "activity")
+            let activity = Activity.fromJSON(parsedActivity) as Activity
 
             var createdAtString = "2014-06-03T00:00:00.000Z"
             var createdAt:NSDate = createdAtString.toNSDate()!