+++ Specs/Model/PostSpec.swift
@@ -13,9 +13,10 @@ class PostSpec: QuickSpec {
     override func spec() {
 
         it("converts from JSON") {            
-            let (parsedPost, parsedLinked) = stubbedJSONDataWithLinked("posts", "posts")
+            let parsedPost = stubbedJSONDataWithLinked("posts", "posts")
+
             let createdAtString = "2014-12-23T22:27:47.325Z"
-            let post = Post.fromJSON(parsedPost, linked: parsedLinked) as Post
+            let post = Post.fromJSON(parsedPost) as Post
             var createdAt:NSDate = createdAtString.toNSDate()!
 
             expect(post.createdAt) == createdAt